+++ src/compiler/linker.c
@@ -54,8 +54,6 @@ static const char *win_subsystem_name(WinSubsystem subsystem)
 {
 	switch (subsystem)
 	{
-		case WIN_SUBSYSTEM_CONSOLE:
-			return "CONSOLE";
 		case WIN_SUBSYSTEM_WINDOWS:
 			return "WINDOWS";
 		case WIN_SUBSYSTEM_NATIVE:
@@ -72,15 +70,15 @@ static const char *win_subsystem_name(WinSubsystem subsystem)
 			return "EFI_ROM";
 		case WIN_SUBSYSTEM_EFI_RUNTIME_DRIVER:
 			return "EFI_RUNTIME_DRIVER";
+		case WIN_SUBSYSTEM_CONSOLE:
 		default:
-			break;
+			return "CONSOLE";
 	}
-	UNREACHABLE
 }
 
 static void linker_setup_windows(const char ***args_ref, Linker linker_type, const char *output_file)
 {
-	if (compiler.build.win.subsystem != WIN_SUBSYSTEM_DEFAULT) add_concat_arg("/SUBSYSTEM:", win_subsystem_name(compiler.build.win.subsystem));
+	add_concat_arg("/SUBSYSTEM:", win_subsystem_name(compiler.build.win.subsystem));
 	if (link_libc()) linking_add_link(&compiler.linking, "dbghelp");
 	if (linker_type == LINKER_CC) return;
 	//add_arg("/MACHINE:X64");