+++ src/compiler/sema_decls.c
@@ -4300,20 +4300,20 @@ static inline bool sema_analyse_main_function(SemaContext *context, Decl *decl)
 		function = decl;
 		goto REGISTER_MAIN;
 	}
-	if (is_win32 && sub_type != MAIN_SUBTYPE_WINMAIN) sub_type = MAIN_SUBTYPE_WMAIN;
-	if (compiler.build.win.subsystem == WIN_SUBSYSTEM_DEFAULT)
+	if (is_win32)
 	{
-		compiler.build.win.subsystem = (sub_type == MAIN_SUBTYPE_WINMAIN) ? WIN_SUBSYSTEM_WINDOWS : WIN_SUBSYSTEM_CONSOLE;
-	}
-	else
-	{
-		if (compiler.build.win.subsystem == WIN_SUBSYSTEM_WINDOWS && sub_type != MAIN_SUBTYPE_WINMAIN)
+		if (sub_type != MAIN_SUBTYPE_WINMAIN) sub_type = MAIN_SUBTYPE_WMAIN;
+		switch (compiler.build.win.subsystem)
 		{
-			RETURN_SEMA_ERROR(decl, "The WINDOWS subsystem requires a WinMain function.");
-		}
-		if (compiler.build.win.subsystem != WIN_SUBSYSTEM_WINDOWS && sub_type == MAIN_SUBTYPE_WINMAIN)
-		{
-			RETURN_SEMA_ERROR(decl, "The WinMain function requires the WINDOWS subsystem.");
+			case WIN_SUBSYSTEM_DEFAULT:
+				compiler.build.win.subsystem = (sub_type == MAIN_SUBTYPE_WINMAIN) ? WIN_SUBSYSTEM_WINDOWS : WIN_SUBSYSTEM_CONSOLE;
+				break;
+			case WIN_SUBSYSTEM_WINDOWS:
+				if (sub_type != MAIN_SUBTYPE_WINMAIN) RETURN_SEMA_ERROR(decl, "The 'windows' subsystem requires a WinMain-style function.");
+				break;
+			default:
+				if (sub_type == MAIN_SUBTYPE_WINMAIN) RETURN_SEMA_ERROR(decl, "A WinMain-style function expected subsystem to be 'windows'.");
+				break;
 		}
 	}
 	function = sema_create_synthetic_main(context, decl, type, sub_type);